+++ example/english.js
@@ -1,11 +1,9 @@
-"use strict";
+const googleTTS = require('..');
 
-var googleTTS = require('..');
-
-googleTTS('Hello World', 'en', 1)   // speed normal = 1 (default), slow = 0.24
-.then(function (url) {
-  console.log(url); // https://translate.google.com/translate_tts?...
-})
-.catch(function (err) {
-  console.error(err.stack);
-});
+googleTTS('Hello World', 'en', 1) // speed normal = 1 (default), slow = 0.24
+  .then((url) => {
+    console.log(url); // https://translate.google.com/translate_tts?...
+  })
+  .catch((err) => {
+    console.error(err.stack);
+  });